Remove GC priorities and improved arms length footer
Removed CSS, template variable and assets related to the GC priority feature. Just kept deprecated CSS to support the old markup of the arms length

Because the arms length footer was defined inside that GC priority block, it was moved outside as an independant feature. It's markup was improved at the same time and it's old markup are now deprecated.