V2 - WIP - Font-size improvement (Default vs Compact)
The idea is to have two type of pages that carry different relative font size for heading and content. Default would be set to 20px and compact would be set to 17px. The compact style would be closer to the font-size used before. Also, the same heading issue for the topic/theme page would be managed separately, where they will be grouped under "top nav". In that approach, the Service initiation font-size change would be in scope of the new size introduced by the "compact" version of the page.